apps/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
openai==1.3.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pillow==10.1.0
numpy==1.25.2
scikit-learn==1.3.2
librosa==0.10.1
whisper-openai==20231117
clip-by-openai==1.0
imagehash==4.3.1
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
supabase==2.0.2
